simx86: asmlinkage should use force_align_arg_pointer attr [fixes #489]
This fixes a cpatch crash that happens when an asmlinkage
function calls to glibc (indirectly, via things like m_munprotect(),
which, with enough log level, goes to *printf stuff):
glibc these days expects the 16-bytes (not bits!) alignment on stack
for SSE (even SSE1) to work with sp-relative locations.
https://gcc.gnu.org/bugzilla/show_bug.cgi?id=40838

This adds the bloat to prologue like this:
	lea    0x8(%rsp),%r10
	and    $0xfffffffffffffff0,%rsp
and to epilogue:
        lea    -0x8(%r10),%rsp
also push/pop'ping r10 in between.